+++ packages/ckeditor5-dev-env/lib/release-tools/templates/header.hbs
