+++ packages/ckeditor5-dev-env/lib/release-tools/utils/getnewreleasetype.js
@@ -5,7 +5,7 @@
 
 'use strict';
 
-const { commitTypes, transform } = require( '../changelog/writer-options' );
+const { commitTypes, transform } = require( './writer-options' );
 
 /**
  * Returns a type (major, minor, patch) of the next release based on commits.
@@ -14,7 +14,7 @@ const { commitTypes, transform } = require( '../changelog/writer-options' );
  */
 module.exports = function getNewReleaseType() {
 	const conventionalRecommendedBump = require( 'conventional-recommended-bump' );
-	const parserOpts = require( '../changelog/parser-options' );
+	const parserOpts = require( './parser-options' );
 
 	return new Promise( ( resolve, reject ) => {
 		const options = {
