+++ packages/ckeditor5-dev-env/tests/release-tools/utils/transformcommit.js
@@ -12,7 +12,7 @@ const sinon = require( 'sinon' );
 const mockery = require( 'mockery' );
 const proxyquire = require( 'proxyquire' );
 
-describe( 'dev-env/release-tools/changelog/writer-options', () => {
+describe( 'dev-env/release-tools/utils/writer-options', () => {
 	describe( 'transform()', () => {
 		let transformCommit, sandbox, stubs, loggerVerbosity;
 
@@ -39,9 +39,9 @@ describe( 'dev-env/release-tools/changelog/writer-options', () => {
 				bugs: 'https://github.com/ckeditor/ckeditor5-dev/issues'
 			} );
 
-			mockery.registerMock( '../utils/getpackagejson', stubs.getPackageJson );
+			mockery.registerMock( './getpackagejson', stubs.getPackageJson );
 
-			transformCommit = proxyquire( '../../../lib/release-tools/changelog/writer-options', {
+			transformCommit = proxyquire( '../../../lib/release-tools/utils/writer-options', {
 				'@ckeditor/ckeditor5-dev-utils': {
 					logger( verbosity ) {
 						loggerVerbosity = verbosity;
